renamed and repositioned AdminUpdateClass to editor folder
