+++ src/pages/Editor/Classes/ClassComponent.styled.tsx
