+++ src/pages/Editor/Classes/UpdateClassModals/AddNewClassModal.tsx
@@ -1,18 +1,18 @@
 import React, { useState } from 'react';
 import { useForm } from 'react-hook-form';
 import { useParams } from 'react-router-dom';
-import { Input, InputField, TextArea } from '../../../components/Input';
-import { ErrorMessageInterface } from '../../../types';
-import { createClass, CreateClassProps } from '../../../api/getPlanClasses';
+import { Input, InputField, TextArea } from '../../../../components/Input';
+import { ErrorMessageInterface } from '../../../../types';
+import { createClass, CreateClassProps } from '../../../../api/getPlanClasses';
 
 import {
   FormErrorMessages,
   StyledErrorMessage,
-} from '../../../components/ErrorMessage';
-import { Form } from '../../../components/Form';
-import { Modal } from '../../../components/Modal/Modal';
-import { Button, PrimaryButton } from '../../../components/Button';
-import { classValidationRules } from '../../../utilities/validation';
+} from '../../../../components/ErrorMessage';
+import { Form } from '../../../../components/Form';
+import { Modal } from '../../../../components/Modal/Modal';
+import { Button, PrimaryButton } from '../../../../components/Button';
+import { classValidationRules } from '../../../../utilities/validation';
 
 const defaultClassValues = {
   name: '',
